requirements.txt
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
numpy==1.18.5
scipy==1.4.1
scikit-learn==0.22
sklearn==0.0
matplotlib==3.3.1
pandas==1.1.0
Pyrebase4==4.3.0